refactor: create helper to get expected Fee events
